PATCH-D: migrate controller and add deprecation shim
