+++ tools/core/bridge/cursor_chatgpt/__init__.py
@@ -1,0 +1,15 @@
+"""Cursor ChatGPT package."""
+
+from .driver import DriverManager
+from .scraper import ChatScraperService
+from .prompt_exec import PromptExecutionService
+from .utils import sanitize_filename
+from .controller import ChatCycleController
+
+__all__ = [
+    "DriverManager",
+    "ChatScraperService",
+    "PromptExecutionService",
+    "sanitize_filename",
+    "ChatCycleController",
+]
